# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Inventory service API in Go using Gin framework",
        "title": "Inventory Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
        "/brand": {
            "get": {
                "description": "Return list of brand.",
                "summary": "Get All brand.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    }
                }
            }
        },
        "/brand/{brandId}": {
            "get": {
                "description": "Return the tahs whoes brandId valu mathes id.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Single brand by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update brand by id",
                        "name": "brandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/golang-crud-gin_pkg_brand_data_response.Response"
                        }
                    }
                }
            }
        },
        "/sku/{skuId}": {
            "get": {
                "description": "Return the tahs whoes skuId valu mathes id.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Single sku by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update sku by id",
                        "name": "skuId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/golang-crud-gin_pkg_sku_data_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "golang-crud-gin_pkg_brand_data_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "golang-crud-gin_pkg_sku_data_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}